+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+### Changed
+
+  - move reusable parts in util directory for actions 
+  - connector docker, stop container, if service not found should not return an error 
+  - deployment enhance exception management and always leave deployment in a final state
   - deployment start, stop, deployment resource changed
 
 ## [0.0.3] - 2019-04-18